read double instead of int

+++ mortageCalc.java
@@ -11,7 +11,7 @@ public class mortageCalc {
         System.out.println("Enter the price of the property : ");
         double propertyPrice = input.nextDouble();
         System.out.println("Enter the rate : ");
-        double rate = input.nextInt();
+        double rate = input.nextDouble();
         System.out.println("How many time a year you make payment : ");
         int freqPayment = input.nextInt();
         System.out.println("How many years are you paying the house? : ");
